+++ react/context/useDeliveryPromise.ts
@@ -11,7 +11,9 @@ import {
   getCatalogCount,
   getPickups,
   updateOrderForm,
+  clearOrderFormShipping,
   updateSession,
+  clearShippingSession,
   getCartProducts,
   orderFormItemsToAvailabilityItems,
   removeCartProductsById,
@@ -644,6 +646,35 @@ export const useDeliveryPromise = () => {
         await updateSession(countryCode, zipcode, geoCoordinates, undefined)
 
         setIsLoading(true)
+        location.reload()
+
+        break
+      }
+
+      case 'CLEAR_ZIPCODE': {
+        const orderFormId = getOrderFormId()
+
+        setIsLoading(true)
+
+        if (orderFormId) {
+          await clearOrderFormShipping(orderFormId)
+        }
+
+        await clearShippingSession()
+
+        setZipCode(undefined)
+        setCity(undefined)
+        setGeoCoordinates(undefined)
+        setPickups([])
+        setSelectedPickup(undefined)
+        setDeliveryPromiseMethod(undefined)
+        setAddressLabel(undefined)
+        setSubmitErrorMessage(undefined)
+        setUnavailableCartItems([])
+        setUnavailabilityMessage(undefined)
+        setActionInterruptedByCartValidation(undefined)
+        setIsLoading(false)
+
         location.reload()
 
         break